requirements.txt
flask
flask-cors
python-dotenv==1.0.0
flask_sqlalchemy
